requirements.txt
requests>=2.31.0
selenium>=4.15.0
pandas>=2.0.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dateutil>=2.8.0
boto3>=1.34.0
botocore>=1.29.0
python-dotenv>=1.0.0
openai>=1.35.0
pinecone-client>=3.0.0